supabase functionality abstracted
