+++ src/app/admin/home/page.jsx
@@ -1,17 +1,12 @@
 "use client";
-import { useState, useEffect } from "react";
-import { createClient } from "@supabase/supabase-js";
+import { useEffect } from "react";
 import { useRouter } from "next/navigation";
-import { FaceIcon, ImageIcon, SunIcon } from "@radix-ui/react-icons";
+import { isLoggedIn } from "@/utils/authentication";
 
 // change icons from svg to radix
 export default function AdminHome() {
   const router = useRouter();
 
-  const supabase = createClient(
-    process.env.NEXT_PUBLIC_SUPABASE_URL,
-    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY
-  );
   function goToLoginError() {
     localStorage.setItem("homeAuthFailed", "true");
     router.push("/admin/login");
@@ -19,15 +14,7 @@ export default function AdminHome() {
 
   useEffect(() => {
     (async () => {
-      const access_token = localStorage.getItem("BCuRm");
-      if (access_token) {
-        const {
-          data: { user },
-        } = await supabase.auth.getUser();
-        if (!user) {
-          goToLoginError();
-        }
-      } else {
+      if (!isLoggedIn()) {
         goToLoginError();
       }
     })();
